+++ _tools/check_docs.ts
@@ -15,6 +15,7 @@ import {
   type DocNodeBase,
   type DocNodeClass,
   type DocNodeFunction,
+  type DocNodeModuleDoc,
   type JsDoc,
   type JsDocTagDocRequired,
   type Location,
@@ -26,23 +27,42 @@ type DocNodeWithJsDoc<T = DocNodeBase> = T & {
 };
 
 const ENTRY_POINTS = [
+  "../assert/mod.ts",
   "../async/mod.ts",
   "../bytes/mod.ts",
   "../cli/mod.ts",
+  "../crypto/mod.ts",
   "../collections/mod.ts",
+  "../csv/mod.ts",
+  "../data_structures/mod.ts",
   "../datetime/mod.ts",
   "../encoding/mod.ts",
+  "../expect/mod.ts",
+  "../fmt/bytes.ts",
+  "../fmt/colors.ts",
+  "../fmt/duration.ts",
+  "../fmt/printf.ts",
+  "../front_matter/mod.ts",
+  "../html/mod.ts",
+  "../http/mod.ts",
   "../internal/mod.ts",
   "../jsonc/mod.ts",
   "../media_types/mod.ts",
+  "../path/mod.ts",
+  "../path/posix/mod.ts",
+  "../path/windows/mod.ts",
+  "../semver/mod.ts",
+  "../streams/mod.ts",
   "../text/mod.ts",
+  "../toml/mod.ts",
   "../ulid/mod.ts",
+  "../url/mod.ts",
+  "../uuid/mod.ts",
   "../webgpu/mod.ts",
-  "../http/mod.ts",
-  "../expect/mod.ts",
 ] as const;
 
 const TS_SNIPPET = /```ts[\s\S]*?```/g;
+const ASSERTION_IMPORT = /import \{.*\} from "@std\/assert(?:\/.*)?";/gm;
 const NEWLINE = "\n";
 const diagnostics: DocumentError[] = [];
 const snippetPromises: Promise<void>[] = [];
@@ -79,6 +99,11 @@ function isVoidOrPromiseVoid(returnType: TsTypeDef) {
       returnType.typeRef.typeName === "Promise" &&
       returnType.typeRef.typeParams?.length === 1 &&
       isVoid(returnType.typeRef.typeParams[0]!));
+}
+
+function isTypeAsserts(returnType: TsTypeDef) {
+  return returnType.kind === "typePredicate" &&
+    returnType.typePredicate.asserts;
 }
 
 function isVoid(returnType: TsTypeDef) {
@@ -122,17 +147,87 @@ function assertHasParamTag(
   }
 }
 
+async function assertSnippetEvals(
+  snippet: string,
+  document: { jsDoc: JsDoc; location: Location },
+) {
+  const command = new Deno.Command(Deno.execPath(), {
+    args: [
+      "eval",
+      "--ext=ts",
+      "--unstable-webgpu",
+      "--no-lock",
+      snippet,
+    ],
+    stderr: "piped",
+  });
+  const timeoutId = setTimeout(() => {
+    console.warn(
+      `Snippet at ${document.location.filename}:${document.location.line} has been running for more than 10 seconds...`,
+    );
+    console.warn(snippet);
+  }, 10_000);
+  try {
+    const { success, stderr } = await command.output();
+    const error = new TextDecoder().decode(stderr);
+    assert(
+      success,
+      `Failed to execute snippet: \n${snippet}\n${error}`,
+      document,
+    );
+  } finally {
+    clearTimeout(timeoutId);
+  }
+}
+
+function assertSnippetsWork(
+  doc: string,
+  document: { jsDoc: JsDoc; location: Location },
+  required = true,
+) {
+  const snippets = doc.match(TS_SNIPPET);
+  if (snippets === null) {
+    if (required) {
+      diagnostics.push(
+        new DocumentError(
+          "@example tag must have a TypeScript code snippet",
+          document,
+        ),
+      );
+      return;
+    } else {
+      return;
+    }
+  }
+  for (let snippet of snippets) {
+    const delim = snippet.split(NEWLINE)[0];
+    if (delim?.includes("no-eval")) continue;
+    // Trim the code block delimiters
+    snippet = snippet.split(NEWLINE).slice(1, -1).join(NEWLINE);
+    if (!delim?.includes("no-assert")) {
+      assert(
+        snippet.match(ASSERTION_IMPORT) !== null,
+        "Snippet must contain assertion from '@std/assert'",
+        document,
+      );
+    }
+    snippetPromises.push(assertSnippetEvals(snippet, document));
+  }
+}
+
 function assertHasExampleTag(
   document: { jsDoc: JsDoc; location: Location },
 ) {
-  const tags = document.jsDoc.tags?.filter((tag) => tag.kind === "example");
+  const tags = document.jsDoc.tags?.filter((tag) =>
+    tag.kind === "example"
+  ) as JsDocTagDocRequired[];
   if (tags === undefined || tags.length === 0) {
     diagnostics.push(
       new DocumentError("Symbol must have an @example tag", document),
     );
     return;
   }
-  for (const tag of (tags as JsDocTagDocRequired[])) {
+  for (const tag of tags) {
     assert(
       tag.doc !== undefined,
       "@example tag must have a title and TypeScript code snippet",
@@ -147,48 +242,7 @@ function assertHasExampleTag(
       "@example tag must have a title",
       document,
     );
-    const snippets = tag.doc.match(TS_SNIPPET);
-    if (snippets === null) {
-      diagnostics.push(
-        new DocumentError(
-          "@example tag must have a TypeScript code snippet",
-          document,
-        ),
-      );
-      continue;
-    }
-    for (let snippet of snippets) {
-      if (snippet.split(NEWLINE)[0]?.includes("no-eval")) continue;
-      // Trim the code block delimiters
-      snippet = snippet.split(NEWLINE).slice(1, -1).join(NEWLINE);
-      const command = new Deno.Command(Deno.execPath(), {
-        args: [
-          "eval",
-          "--ext=ts",
-          "--unstable-webgpu",
-          snippet,
-        ],
-        stderr: "piped",
-      });
-      snippetPromises.push((async () => {
-        const timeoutId = setTimeout(() => {
-          console.warn("Snippet has been running for more than 10 seconds...");
-          console.warn(snippet);
-        }, 10_000);
-        try {
-          const { success, stderr } = await command.output();
-          assert(
-            success,
-            `Example code snippet failed to execute: \n${snippet}\n${
-              new TextDecoder().decode(stderr)
-            }`,
-            document,
-          );
-        } finally {
-          clearTimeout(timeoutId);
-        }
-      })());
-    }
+    assertSnippetsWork(tag.doc, document);
   }
 }
 
@@ -227,6 +281,7 @@ function assertHasTypeParamTags(
 function assertFunctionDocs(
   document: DocNodeWithJsDoc<DocNodeFunction | ClassMethodDef>,
 ) {
+  assertSnippetsWork(document.jsDoc.doc!, document, false);
   for (const param of document.functionDef.params) {
     if (param.kind === "identifier") {
       assertHasParamTag(document, param.name);
@@ -240,7 +295,8 @@ function assertFunctionDocs(
   }
   if (
     document.functionDef.returnType !== undefined &&
-    !isVoidOrPromiseVoid(document.functionDef.returnType)
+    !isVoidOrPromiseVoid(document.functionDef.returnType) &&
+    !isTypeAsserts(document.functionDef.returnType)
   ) {
     assertHasReturnTag(document);
   }
@@ -255,6 +311,7 @@ function assertFunctionDocs(
  * - Documentation on all properties, methods, and constructors.
  */
 function assertClassDocs(document: DocNodeWithJsDoc<DocNodeClass>) {
+  assertSnippetsWork(document.jsDoc.doc!, document, false);
   for (const typeParam of document.classDef.typeParams) {
     assertHasTypeParamTags(document, typeParam.name);
   }
@@ -340,12 +397,31 @@ function assertConstructorDocs(
   assertHasExampleTag(constructor);
 }
 
+/**
+ * Checks a module document for:
+ * - Code snippets that execute successfully.
+ */
+function assertModuleDoc(document: DocNodeWithJsDoc<DocNodeModuleDoc>) {
+  assertSnippetsWork(document.jsDoc.doc!, document);
+}
+
+function resolve(specifier: string, referrer: string): string {
+  if (specifier.startsWith("@std/") && specifier.split("/").length > 2) {
+    specifier = specifier.replace("@std/", "../").replaceAll("-", "_") + ".ts";
+  }
+  return new URL(specifier, referrer).href;
+}
+
 async function checkDocs(specifier: string) {
-  const docs = await doc(specifier);
+  const docs = await doc(specifier, { resolve });
   for (const d of docs.filter(isExported)) {
     if (d.jsDoc === undefined) continue; // this is caught by other checks
     const document = d as DocNodeWithJsDoc<DocNode>;
     switch (document.kind) {
+      case "moduleDoc": {
+        assertModuleDoc(document);
+        break;
+      }
       case "function": {
         assertFunctionDocs(document);
         break;
@@ -393,5 +469,7 @@ if (diagnostics.length > 0) {
       "color: gray",
     );
   }
+
+  console.log(`%c${diagnostics.length} errors found`, "color: red");
   Deno.exit(1);
 }