+++ assert/assert_string_includes.ts
@@ -6,22 +6,25 @@ import { AssertionError } from "./assertion_error.ts";
  * Make an assertion that actual includes expected. If not
  * then throw.
  *
- * @example
- * ```ts
+ * @example Usage
+ * ```ts no-eval
  * import { assertStringIncludes } from "@std/assert/assert-string-includes";
  *
  * assertStringIncludes("Hello", "ello"); // Doesn't throw
  * assertStringIncludes("Hello", "world"); // Throws
  * ```
+ *
+ * @param actual The actual string to check for inclusion.
+ * @param expected The expected string to check for inclusion.
+ * @param msg The optional message to display if the assertion fails.
  */
 export function assertStringIncludes(
   actual: string,
   expected: string,
   msg?: string,
 ) {
-  if (!actual.includes(expected)) {
-    const msgSuffix = msg ? `: ${msg}` : ".";
-    msg = `Expected actual: "${actual}" to contain: "${expected}"${msgSuffix}`;
-    throw new AssertionError(msg);
-  }
+  if (actual.includes(expected)) return;
+  const msgSuffix = msg ? `: ${msg}` : ".";
+  msg = `Expected actual: "${actual}" to contain: "${expected}"${msgSuffix}`;
+  throw new AssertionError(msg);
 }