+++ encoding/base58.ts
@@ -6,8 +6,6 @@
  * {@link https://datatracker.ietf.org/doc/html/draft-msporny-base58-03 | base58}
  * encoding and decoding.
  *
- * This module is browser compatible.
- *
  * ```ts
  * import { encodeBase58, decodeBase58 } from "@std/encoding/base58";
  * import { assertEquals } from "@std/assert/assert-equals";
@@ -22,7 +20,7 @@
  * @module
  */
 
-import { validateBinaryLike } from "./_util.ts";
+import { validateBinaryLike } from "./_validate_binary_like.ts";
 
 // deno-fmt-ignore
 const mapBase58: Record<string, number> = {
@@ -145,7 +143,9 @@ export function decodeBase58(b58: string): Uint8Array {
     let i = 0;
 
     if (carry === undefined) {
-      throw new Error(`Invalid base58 char at index ${idx} with value ${char}`);
+      throw new TypeError(
+        `Invalid base58 char at index ${idx} with value ${char}`,
+      );
     }
 
     for (