+++ path/common_test.ts
@@ -1,17 +1,18 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { assertEquals } from "@std/assert";
 import { common } from "./mod.ts";
+import * as posix from "./posix/mod.ts";
+import * as windows from "./windows/mod.ts";
 
 Deno.test({
   name: "common() returns shared path",
   fn() {
-    const actual = common(
+    const actual = posix.common(
       [
         "file://deno/cli/js/deno.ts",
         "file://deno/std/path/mod.ts",
         "file://deno/cli/js/main.ts",
       ],
-      "/",
     );
     assertEquals(actual, "file://deno/");
   },
@@ -20,9 +21,8 @@ Deno.test({
 Deno.test({
   name: "common() returns empty string if no shared path is present",
   fn() {
-    const actual = common(
+    const actual = posix.common(
       ["file://deno/cli/js/deno.ts", "https://deno.land/std/path/mod.ts"],
-      "/",
     );
     assertEquals(actual, "");
   },
@@ -31,13 +31,12 @@ Deno.test({
 Deno.test({
   name: "common() checks windows separator",
   fn() {
-    const actual = common(
+    const actual = windows.common(
       [
         "c:\\deno\\cli\\js\\deno.ts",
         "c:\\deno\\std\\path\\mod.ts",
         "c:\\deno\\cli\\js\\main.ts",
       ],
-      "\\",
     );
     assertEquals(actual, "c:\\deno\\");
   },
@@ -46,7 +45,7 @@ Deno.test({
 Deno.test({
   name: "common(['', '/'], '/') returns ''",
   fn() {
-    const actual = common(["", "/"], "/");
+    const actual = posix.common(["", "/"]);
     assertEquals(actual, "");
   },
 });
@@ -54,10 +53,10 @@ Deno.test({
 Deno.test({
   name: "common(['/', ''], '/') returns ''",
   fn() {
-    const actual = common([
+    const actual = posix.common([
       "/",
       "",
-    ], "/");
+    ]);
     assertEquals(actual, "");
   },
 });
@@ -65,7 +64,7 @@ Deno.test({
 Deno.test({
   name: "common() returns the first path unmodified when it's the only path",
   fn() {
-    const actual = common(["./deno/std/path/mod.ts"], "/");
+    const actual = posix.common(["./deno/std/path/mod.ts"]);
     assertEquals(actual, "./deno/std/path/mod.ts");
   },
 });
@@ -79,8 +78,21 @@ Deno.test({
         "./deno/std/path/mod.ts",
         "./deno/std/path/mod.ts",
       ],
-      "/",
     );
     assertEquals(actual, "./deno/std/path/mod.ts");
+  },
+});
+
+Deno.test({
+  name: "posix.common() returns shared path",
+  fn() {
+    const actual = posix.common(
+      [
+        "file://deno/cli/js/deno.ts",
+        "file://deno/std/path/mod.ts",
+        "file://deno/cli/js/main.ts",
+      ],
+    );
+    assertEquals(actual, "file://deno/");
   },
 });