+++ path/posix/relative.ts
@@ -8,8 +8,20 @@ import { assertArgs } from "../_common/relative.ts";
 /**
  * Return the relative path from `from` to `to` based on current working directory.
  *
- * @param from path in current working directory
- * @param to path in current working directory
+ * If `from` and `to` are the same, return an empty string.
+ *
+ * @example Usage
+ * ```ts
+ * import { relative } from "@std/path/posix/relative";
+ * import { assertEquals } from "@std/assert/assert-equals";
+ *
+ * const path = relative("/data/orandea/test/aaa", "/data/orandea/impl/bbb");
+ * assertEquals(path, "../../impl/bbb");
+ * ```
+ *
+ * @param from The path to start from.
+ * @param to The path to reach.
+ * @returns The relative path.
  */
 export function relative(from: string, to: string): string {
   assertArgs(from, to);