+++ text/word_similarity_sort.ts
@@ -3,26 +3,27 @@
 import { compareSimilarity } from "./compare_similarity.ts";
 
 /**
- * Sorts a string-array by similarity to a given string
+ * Sorts a string-array by similarity to a given string.
  *
- * @example Usage
+ * @example Basic usage
+ *
+ * This function is case-insensitive by default.
+ *
  * ```ts
  * import { wordSimilaritySort } from "@std/text/word-similarity-sort";
+ * import { assertEquals } from "@std/assert/assert-equals";
  *
  * const possibleWords = ["length", "size", "blah", "help"];
- *
- * // case-insensitive by default
  * const suggestions = wordSimilaritySort("hep", possibleWords).join(", ");
  *
- * // force case sensitive
- * wordSimilaritySort("hep", possibleWords, { caseSensitive: true });
+ * assertEquals(suggestions, "help, size, blah, length");
  * ```
  *
  * @param givenWord - The string to measure distance against
  * @param possibleWords - The string-array that will be sorted
  * @param options An options bag containing a `caseSensitive` flag indicating
  * whether the distance should include case. Default is false.
- * @returns {string[]} A sorted copy of possibleWords
+ * @returns A sorted copy of possibleWords
  */
 export function wordSimilaritySort(
   givenWord: string,